requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.2
coverage==4.3.4
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
flake8==3.5.0
future==0.15.2
graphviz==0.8.1
html5lib==1.0.1
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.12
jupyterlab-launcher==0.10.5
linecache2==1.0.0
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
nose-exclude==0.5.0
notebook==5.2.2
numpy==1.16.0
ordereddict==1.1
pandas==0.21.1
pandas-ml==0.5.0
pandocfilters==1.4.2
parso==0.1.0
pbr==4.0.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psutil==5.4.3
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
scikit-learn==0.19.1
scipy==1.2.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traceback2==1.4.0
traitlets==4.3.2
unittest2==1.1.0
waitress==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.7.4
WebTest==2.0.29
widgetsnbextension==3.0.8
xgboost==0.81
xlrd==1.1.0
xlwt==1.3.0
